# backend/api/gen/go/recommend/service/v1/recommend_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.0
// - protoc             (unknown)
// source: recommend/service/v1/recommend.proto

package servicev1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationRecommendServiceGetRecommendHistory = "/recommend.service.v1.RecommendService/GetRecommendHistory"
const OperationRecommendServiceTriggerRecommend = "/recommend.service.v1.RecommendService/TriggerRecommend"

type RecommendServiceHTTPServer interface {
	GetRecommendHistory(context.Context, *RecommendHistoryRequest) (*RecommendHistoryResponse, error)
	TriggerRecommend(context.Context, *RecommendRequest) (*RecommendResponse, error)
}

func RegisterRecommendServiceHTTPServer(s *http.Server, srv RecommendServiceHTTPServer) {
	r := s.Route("/")
	r.POST("/api/v1/recommend", _RecommendService_TriggerRecommend0_HTTP_Handler(srv))
	r.GET("/api/v1/recommend/history", _RecommendService_GetRecommendHistory0_HTTP_Handler(srv))
}

func _RecommendService_TriggerRecommend0_HTTP_Handler(srv RecommendServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in RecommendRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRecommendServiceTriggerRecommend)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.TriggerRecommend(ctx, req.(*RecommendRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*RecommendResponse)
		return ctx.Result(200, reply)
	}
}

func _RecommendService_GetRecommendHistory0_HTTP_Handler(srv RecommendServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in RecommendHistoryRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRecommendServiceGetRecommendHistory)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetRecommendHistory(ctx, req.(*RecommendHistoryRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*RecommendHistoryResponse)
		return ctx.Result(200, reply)
	}
}

type RecommendServiceHTTPClient interface {
	GetRecommendHistory(ctx context.Context, req *RecommendHistoryRequest, opts ...http.CallOption) (rsp *RecommendHistoryResponse, err error)
	TriggerRecommend(ctx context.Context, req *RecommendRequest, opts ...http.CallOption) (rsp *RecommendResponse, err error)
}

type RecommendServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewRecommendServiceHTTPClient(client *http.Client) RecommendServiceHTTPClient {
	return &RecommendServiceHTTPClientImpl{client}
}

func (c *RecommendServiceHTTPClientImpl) GetRecommendHistory(ctx context.Context, in *RecommendHistoryRequest, opts ...http.CallOption) (*RecommendHistoryResponse, error) {
	var out RecommendHistoryResponse
	pattern := "/api/v1/recommend/history"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationRecommendServiceGetRecommendHistory))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *RecommendServiceHTTPClientImpl) TriggerRecommend(ctx context.Context, in *RecommendRequest, opts ...http.CallOption) (*RecommendResponse, error) {
	var out RecommendResponse
	pattern := "/api/v1/recommend"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationRecommendServiceTriggerRecommend))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}